+++ exercicios/src/CalculadoraSubtracao.java
@@ -1,6 +1,6 @@
 public class CalculadoraSubtracao {
-    public int main(int num1,int num2) {
-        int subtracao = num1 - num2;
+    public double main(double num1,double num2) {
+        double subtracao = num1 - num2;
         return subtracao;
     }
 }
